k-food-helpdesk-claude/server/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
google-genai>=1.0.0
openai>=1.40.0
psycopg2-binary>=2.9.9
PyJWT>=2.9.0
pytest>=8.3.0
python-dotenv>=1.0.1
redis>=5.0.0
sentence-transformers>=3.0.0
uvicorn[standard]>=0.30.0
